微调 POS应收int->double V2.0 doc完成

+++ spt/src/main/java/com/hit/spt/service/OrderService.java
@@ -15,7 +15,7 @@ public interface OrderService {
 
     List<OrderItem> queryOrderItemWithNameListByOid(Integer o_id);
 
-    int calcTotalPriceByOid(Integer o_id);
+    Double calcTotalPriceByOid(Integer o_id);
 
     OrderItem generateOrderItem(Integer o_id, String name, Integer quantity, Boolean trade);
